Updated Builder to reuse same CodeGeneratorConfig
