+++ hollow/src/main/java/com/netflix/hollow/api/codegen/api/HollowTypeAPIGenerator.java
@@ -17,7 +17,7 @@ package com.netflix.hollow.api.codegen.api;
 
 import static com.netflix.hollow.api.codegen.HollowCodeGenerationUtils.typeAPIClassname;
 
-import com.netflix.hollow.api.codegen.HollowAPIGenerator.CodeGeneratorConfig;
+import com.netflix.hollow.api.codegen.CodeGeneratorConfig;
 import com.netflix.hollow.api.codegen.HollowConsumerJavaFileGenerator;
 import com.netflix.hollow.core.schema.HollowSchema;
