+++ hollow/src/main/java/com/netflix/hollow/api/codegen/delegate/HollowObjectDelegateGenerator.java
@@ -15,7 +15,7 @@
  */
 package com.netflix.hollow.api.codegen.delegate;
 
-import com.netflix.hollow.api.codegen.HollowAPIGenerator.CodeGeneratorConfig;
+import com.netflix.hollow.api.codegen.CodeGeneratorConfig;
 import com.netflix.hollow.api.codegen.HollowConsumerJavaFileGenerator;
 import com.netflix.hollow.api.codegen.HollowErgonomicAPIShortcuts;
 import com.netflix.hollow.core.schema.HollowObjectSchema;
