+++ hollow/src/main/java/com/netflix/hollow/api/codegen/indexes/LegacyHollowPrimaryKeyIndexGenerator.java
@@ -17,8 +17,8 @@
  */
 package com.netflix.hollow.api.codegen.indexes;
 
+import com.netflix.hollow.api.codegen.CodeGeneratorConfig;
 import com.netflix.hollow.api.codegen.HollowAPIGenerator;
-import com.netflix.hollow.api.codegen.HollowAPIGenerator.CodeGeneratorConfig;
 import com.netflix.hollow.api.custom.HollowAPI;
 import com.netflix.hollow.core.schema.HollowObjectSchema;
