+++ hollow/src/main/java/com/netflix/hollow/api/codegen/objects/HollowCollectionsGenerator.java
@@ -15,7 +15,7 @@
  */
 package com.netflix.hollow.api.codegen.objects;
 
-import com.netflix.hollow.api.codegen.HollowAPIGenerator.CodeGeneratorConfig;
+import com.netflix.hollow.api.codegen.CodeGeneratorConfig;
 import com.netflix.hollow.api.codegen.HollowConsumerJavaFileGenerator;
 import com.netflix.hollow.core.schema.HollowSchema;
